Fixed negative score in Apple game. Now sets score to 0 if it tries to go negative.
Removed the default font and added a custom font for the apple game and
custom font for the menu if needed.

Commented out the setScale for the fonts because it stretched the font
size and made it look pixelated.

Still need to change font color, because white doesn't stick out very
well in the game. Perhaps Red or Green?